Update move import to own router
